requirements.txt
asgiref==3.8.1
chardet==5.2.0
dj-database-url==2.3.0
Django==5.1.3
django-environ==0.11.2
django-rest-framework==0.1.0
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
PyPDF2==3.0.1
python-decouple==3.8
qrcode==8.0
reportlab==4.2.5
sqlparse==0.5.2
typing_extensions==4.12.2
whitenoise==6.8.2
